refactor stat viewer elements into separate files
