+++ views/digitaljs/src/viewers/index.ts
@@ -1,3 +1,3 @@
-export * from './diagram/diagram';
-export * from './stats/stats';
-export * from './base';
+export {DiagramViewer} from './diagram';
+export {StatsViewer} from './stats';
+export {BaseViewer} from './base';
