+++ views/digitaljs/src/viewers/stats/elements/index.ts
@@ -1,0 +1,4 @@
+export {ModuleOverviewGrid} from './moduleoverview';
+export {TabsContainer} from './tabs';
+export {ModuleExplorerGrid} from './explorer';
+export {PrimitivesOverviewGrid} from './primitives';
